Updated AddressRequest.kt

+++ src/main/kotlin/br/com/tcc/petsus/domain/model/api/address/request/AddressRequest.kt
@@ -12,7 +12,7 @@ import javax.validation.constraints.Size
 data class AddressRequest(
     @SerializedName("lat") @NotNull val lat: Double,
     @SerializedName("lng") @NotNull val lng: Double,
-    @SerializedName("number") @NotNull var number: Int,
+    @SerializedName("number") @NotNull var number: Long,
     @SerializedName("cityId") @NotNull val cityId: Long? = null,
     @SerializedName("complement") var complement: String? = null,
     @SerializedName("address") @NotNull @NotBlank var address: String,
